feat: Add .NET support for project build and execution
The code changes introduce support for building and executing .NET projects. The `try_dotnet()` function is added to detect the presence of .csproj, .fsproj, or .sln files and execute the appropriate dotnet command. This allows users to build, run, and test C# and F# projects using the `dotnet build`, `dotnet run`, and `dotnet test` commands respectively.