feat: ✨ introduce support for pnpm
- Implemented detection of pnpm by checking for the presence of the
`pnpm-lock.yaml` file.

- Maintained consistency with the existing pattern of package
manager detection (npm and yarn)

Note: While implementing this feature, it's worth mentioning
that the `packageManager` directive in the `package.json` file
an be used to specify the package manager for the project.
This provides a more standardized approach to identifying the package manager.

For more details, refer to the official documentation
[here](https://nodejs.org/api/packages.html#packagemanager).